1、整理了程序的空间分布，但还没有完全测试

+++ src/boot_param.h
@@ -29,7 +29,7 @@ typedef struct
     uint32_t magic;//魔术字，用于参数头部校验
     uint32_t lenth;//本参数结构体的长度
     
-    uint8_t version;    //bootloader`
+    uint32_t version;    //bootloader`
     uint8_t debug_mode; //调试模式
     uint8_t wait_sec;   //等待键盘事件秒数
     uint8_t run_type;   //运行在RAM或Flash
